requirements.txt
pyautogen
openai
python-dotenv